+++ draft-packages/menu/KaizenDraft/Menu/MenuDropdown.tsx
@@ -1,8 +1,10 @@
+import classnames from "classnames"
 import * as React from "react"
 
 const styles = require("./styles.scss")
 
 type MenuDropdownProps = {
+  id?: string
   hideMenuDropdown: () => void
   position?: {
     top: number
@@ -10,6 +12,8 @@ type MenuDropdownProps = {
     left: number
     right: number
   } | null
+  align?: "left" | "right"
+  width?: "default" | "contain"
 }
 
 export default class MenuDropdown extends React.Component<MenuDropdownProps> {
@@ -39,13 +43,8 @@ export default class MenuDropdown extends React.Component<MenuDropdownProps> {
       const { innerHeight } = window
       const rect = menu.current.getBoundingClientRect()
 
-      if (pos.bottom > innerHeight - rect.height) {
-        menu.current.style.bottom = "24px"
-        menu.current.style.top = "auto"
-      } else {
-        menu.current.style.top = "24px"
-        menu.current.style.bottom = "auto"
-      }
+      menu.current.style.bottom =
+        pos.bottom > innerHeight - rect.height ? "24px" : "auto"
     }
   }
 
@@ -65,11 +64,20 @@ export default class MenuDropdown extends React.Component<MenuDropdownProps> {
   }
 
   render(): JSX.Element {
-    const { hideMenuDropdown, children } = this.props
+    const {
+      hideMenuDropdown,
+      children,
+      align = "left",
+      width = "default",
+    } = this.props
 
     return (
       <div
-        className={styles.menuContainer}
+        id={this.props.id}
+        className={classnames(styles.menuContainer, {
+          [styles.defaultWidth]: width == "default",
+          [styles.alignRight]: align == "right",
+        })}
         ref={this.menu}
         onClick={() => hideMenuDropdown()}
       >