+++ draft-packages/menu/CHANGELOG.md
@@ -3,6 +3,84 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.2.1...@kaizen/draft-menu@2.3.0) (2020-07-07)
+
+
+### Features
+
+* add `id` prop to MenuDropdown and Menu ([#620](https://github.com/cultureamp/kaizen-design-system/issues/620)) ([7e33143](https://github.com/cultureamp/kaizen-design-system/commit/7e331433d6faf114dd85e4348edb1aaa072dd18a))
+
+
+
+
+
+## [2.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.2.0...@kaizen/draft-menu@2.2.1) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+# [2.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.1.0...@kaizen/draft-menu@2.2.0) (2020-07-05)
+
+
+### Features
+
+* add dropdownWidth prop to Menu ([#613](https://github.com/cultureamp/kaizen-design-system/issues/613)) ([c782a43](https://github.com/cultureamp/kaizen-design-system/commit/c782a4359d02cb87f996b5f1448414c1ae11ac1b))
+
+
+
+
+
+# [2.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@2.0.0...@kaizen/draft-menu@2.1.0) (2020-07-03)
+
+
+### Features
+
+* StatelessMenu ([#607](https://github.com/cultureamp/kaizen-design-system/issues/607)) ([00fe797](https://github.com/cultureamp/kaizen-design-system/commit/00fe797fc3f9fefebc8f4a0945d40c3ea1bd6ea1))
+
+
+
+
+
+# [2.0.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@1.2.4...@kaizen/draft-menu@2.0.0) (2020-07-02)
+
+
+### Features
+
+* Menu is now left aligned by default ([#605](https://github.com/cultureamp/kaizen-design-system/issues/605)) ([fc29a2f](https://github.com/cultureamp/kaizen-design-system/commit/fc29a2f321a27f62273f763bc53667b67365fcc5))
+
+
+### BREAKING CHANGES
+
+* Menu is now left aligned by default
+
+adds an `align` prop to Menu and MenuDropdown with
+options `"left"` and `"right"`. Previously righ-aligned
+was the only option. Now left aligned is an option and
+it becomes the default. Consumers must add `align="right"`
+to any existing usages to retain identical functionality.
+
+* feat: further Zenify the MenuDropdown component
+
+- the dropdown is now flush to the right, and
+  apears below the button with a small gap
+
+* docs: add story for edge case where button is at bottom of screen
+
+
+
+
+
+## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@1.2.3...@kaizen/draft-menu@1.2.4) (2020-07-01)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
 ## [1.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@1.2.2...@kaizen/draft-menu@1.2.3) (2020-06-29)
 
 **Note:** Version bump only for package @kaizen/draft-menu