+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.24...@kaizen/draft-table@1.6.25) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.23...@kaizen/draft-table@1.6.24) (2020-07-01)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
 ## [1.6.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.22...@kaizen/draft-table@1.6.23) (2020-06-29)
 
 **Note:** Version bump only for package @kaizen/draft-table