+++ draft-packages/title-block/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.19...@kaizen/draft-title-block@1.5.20) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [1.5.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.18...@kaizen/draft-title-block@1.5.19) (2020-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [1.5.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.17...@kaizen/draft-title-block@1.5.18) (2020-07-01)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
 ## [1.5.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.16...@kaizen/draft-title-block@1.5.17) (2020-06-29)