+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.24.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.24.1...@kaizen/component-library@7.24.2) (2020-07-06)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
+## [7.24.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.24.0...@kaizen/component-library@7.24.1) (2020-07-01)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 # [7.24.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.23.7...@kaizen/component-library@7.24.0) (2020-06-29)